requirements.txt
uvicorn==0.11.8
python-multipart==0.0.5
scipy==1.5.2
imgaug==0.4.0
numpy==1.19.1
opencv-python-headless==4.4.0.42
Pillow==6.2.0
Cython==0.29.10
pandas==1.1.2
torch
torchvision
opencv-python
multiprocess
easydict
PyYAML
gc-python-utils
watchdog